+++ blocklets/snap-kit/api/src/libs/env.ts
@@ -1,5 +1,6 @@
 import env from '@blocklet/sdk/lib/env';
 
 export default {
+  isDev: process.env.NODE_ENV === 'development',
   ...env,
 };
