+++ blocklets/snap-kit/blocklet.prefs.json
@@ -1,0 +1,145 @@
+{
+  "form": {
+    "labelCol": 6,
+    "wrapperCol": 12
+  },
+  "schema": {
+    "type": "object",
+    "properties": {
+      "siteCron": {
+        "type": "array",
+        "x-decorator": "FormItem",
+        "x-component": "ArrayTable",
+        "x-validator": [],
+        "x-component-props": {},
+        "x-decorator-props": {},
+        "name": "siteCron",
+        "title": "",
+        "x-designable-id": "62hw9y9yg1j",
+        "x-index": 0,
+        "items": {
+          "type": "object",
+          "x-designable-id": "ujexcbb8cph",
+          "properties": {
+            "a89gnk1caxv": {
+              "type": "void",
+              "x-component": "ArrayTable.Column",
+              "x-component-props": {
+                "title": "Site Url"
+              },
+              "name": "",
+              "x-designable-id": "a89gnk1caxv",
+              "x-index": 0,
+              "properties": {
+                "url": {
+                  "type": "string",
+                  "title": "",
+                  "x-decorator": "FormItem",
+                  "x-component": "Input",
+                  "x-validator": [],
+                  "x-component-props": {},
+                  "x-decorator-props": {},
+                  "name": "url",
+                  "required": true,
+                  "x-designable-id": "tnjie4lllmx",
+                  "x-index": 0
+                }
+              }
+            },
+            "qgiensr1fix": {
+              "type": "void",
+              "x-component": "ArrayTable.Column",
+              "x-component-props": {
+                "title": "Pathname"
+              },
+              "name": "",
+              "x-designable-id": "qgiensr1fix",
+              "x-index": 1,
+              "properties": {
+                "pathname": {
+                  "type": "string",
+                  "title": "",
+                  "x-decorator": "FormItem",
+                  "x-component": "Input",
+                  "x-validator": [],
+                  "x-component-props": {},
+                  "x-decorator-props": {},
+                  "default": "/",
+                  "name": "pathname",
+                  "required": true,
+                  "x-designable-id": "zeg6n47ghxt",
+                  "x-index": 0
+                }
+              }
+            },
+            "w8kld6v7gf0": {
+              "type": "void",
+              "x-component": "ArrayTable.Column",
+              "x-component-props": {
+                "title": "Interval"
+              },
+              "name": "",
+              "x-designable-id": "w8kld6v7gf0",
+              "x-index": 2,
+              "properties": {
+                "interval": {
+                  "type": "string",
+                  "x-decorator": "FormItem",
+                  "x-component": "Input",
+                  "x-validator": [],
+                  "x-component-props": {},
+                  "x-decorator-props": {},
+                  "name": "interval",
+                  "default": "0",
+                  "description": "",
+                  "x-designable-id": "n7umbeejnu5",
+                  "x-index": 0
+                }
+              }
+            },
+            "jryok60h0oy": {
+              "type": "void",
+              "x-component": "ArrayTable.Column",
+              "x-component-props": {
+                "title": "Title"
+              },
+              "x-designable-id": "jryok60h0oy",
+              "x-index": 3,
+              "properties": {
+                "z0zvrclj14h": {
+                  "type": "void",
+                  "x-component": "ArrayTable.Remove",
+                  "x-designable-id": "z0zvrclj14h",
+                  "x-index": 0
+                },
+                "bmeqc7itlg0": {
+                  "type": "void",
+                  "x-component": "ArrayTable.MoveDown",
+                  "x-designable-id": "bmeqc7itlg0",
+                  "x-index": 1
+                },
+                "lzobd4rqfuv": {
+                  "type": "void",
+                  "x-component": "ArrayTable.MoveUp",
+                  "x-designable-id": "lzobd4rqfuv",
+                  "x-index": 2
+                }
+              }
+            }
+          }
+        },
+        "properties": {
+          "10c6ljjp047": {
+            "type": "void",
+            "title": "Addition",
+            "x-component": "ArrayTable.Addition",
+            "x-component-props": {},
+            "x-designable-id": "10c6ljjp047",
+            "x-index": 0
+          }
+        }
+      }
+    },
+    "x-designable-id": "4wk51ka49fc"
+  }
+}
