+++ blocklets/snap-kit/package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "@abtnode/cron": "^1.16.43",
     "@arcblock/crawler": "workspace:*",
+    "@arcblock/crawler-middleware": "workspace:*",
     "@arcblock/did-auth": "^1.19.15",
     "@arcblock/did-auth-storage-nedb": "^1.7.1",
     "@arcblock/validator": "^1.20.11",
