+++ packages/crawler/src/store/index.ts
@@ -3,20 +3,26 @@ import { SqliteDialect } from '@sequelize/sqlite3';
 import path from 'path';
 
 import { config, logger } from '../config';
-import { initJobModel } from './job';
-import { initSnapshotModel } from './snapshot';
+import { Job } from './job';
+import { Snapshot } from './snapshot';
 
-export async function ensureDatabase() {
+export async function initDatabase() {
   const sequelize = new Sequelize({
     dialect: SqliteDialect,
     storage: path.join(config.dataDir, 'snap-kit.db'),
     logging: (msg) => process.env.SQLITE_LOG && logger.debug(msg),
   });
 
-  await initSnapshotModel(sequelize);
-  await initJobModel(sequelize);
+  Job.initModel(sequelize);
+  Snapshot.initModel(sequelize);
 
   try {
+    await Promise.all([
+      sequelize.query('pragma journal_mode = WAL;'),
+      sequelize.query('pragma synchronous = normal;'),
+      sequelize.query('pragma journal_size_limit = 67108864;'),
+    ]);
+
     await sequelize.authenticate();
     await sequelize.sync();
     logger.info('Successfully connected to database');
