+++ packages/middleware/README.md
@@ -1,0 +1,49 @@
+# @arcblock/crawler-middleware
+
+This express middleware provides pre-rendered HTML generated by SnapKit for Blocklets, enabling them to return complete HTML content to web crawlers. This is essential for SEO and ensuring that search engines can properly index dynamically generated content.
+
+## Usage
+
+```typescript
+import { createSnapshotMiddleware } from '@arcblock/crawler-middleware';
+
+const app = express();
+
+app.use(
+  createSnapshotMiddleware({
+    endpoint: process.env.SNAP_KIT_ENDPOINT,
+    accessKey: process.env.SNAP_KIT_ACCESS_KEY,
+    allowCrawler: (req) => {
+      return req.path === '/';
+    },
+  }),
+);
+```
+
+## How it Works
+
+1.  The middleware intercepts incoming requests.
+2.  It checks if the request is from a web crawler.
+3.  Try to read and return HTML from the local cache.
+4.  If the cache is not found, an asynchronous request is made to SnapKit, and the local cache is updated.
+5.  The current request does not return the cached content; the next crawler visit will hit step 3 and return the cache directly.
+
+## Options
+
+The options for createSnapshotMiddleware:
+
+```typescript
+{
+  /** SnapKit endpoint */
+  endpoint: string;
+  /** SnapKit access key */
+  accessKey: string;
+  /** Max cache size for LRU cache */
+  cacheMax?: number;
+  /**
+   * Cache update interval
+   * When cache exceeds this time, it will try to fetch and update cache from SnapKit
+   */
+  cacheUpdateInterval?: number;
+};
+```
