+++ packages/middleware/tsconfig.cjs.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "tests"],
+  "compilerOptions": {
+    "module": "commonjs",
+    "outDir": "dist"
+  }
+}
