+++ packages/middleware/tsconfig.esm.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src"],
+  "exclude": ["node_modules", "tests"],
+  "compilerOptions": {
+    "module": "esnext",
+    "outDir": "esm",
+    "moduleResolution": "node"
+  }
+}
